Added in a second ship for reliable-er collision testing, took out hte goddamn triangle. Todo: init polygons easierly, render boxes in a function, and with a flag like dev-debug mode

+++ PolygonCollisions/Collisions.m
@@ -29,7 +29,7 @@
     
     int edgeCountA = polygonA.pointCount;
     int edgeCountB = polygonB.pointCount;
-    float  minIntervalDistance = DBL_MAX; // Should be infinity though
+    float  minIntervalDistance = FLT_MAX; // Should be infinity though
     Vector2f translationAxis = Vector2fZero;
     Vector2f edge;
